+++ JavaFXApplication1/src/main/java/com/mycompany/javafxapplication1/controller/UpdateFileController.java
@@ -78,7 +78,7 @@ public class UpdateFileController {
         }
     }
     
-    private void loadContentAsync(long fileId) {
+    private void loadContentAsync(int fileId) {
         new Thread(() -> {
             try {
                 String reqId = fileService.loadContent(fileId);
